Add wrapper class for PyPI interaction
